backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.2
websockets==13.0
alpaca-py==0.28.1
ib_insync==0.9.86
polygon-api-client==1.14.0
praw==7.7.1
airtable-python-wrapper==0.15.3
supabase==2.7.0
apscheduler==3.10.4
python-telegram-bot==21.5
twilio==9.2.0
sendgrid==6.11.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
ta==0.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
